SalomonAI_Organizado/services/simulation-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
pydantic-settings==2.3.3
python-jose[cryptography]==3.3.0
httpx==0.27.2
